Move SyncBuffer to integration test package
